mmr to mf6dfw test complete with boundary conditions
